+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Unreleased
 
+- Add support for `AsyncIterator`, `io.Reader`, `io.Writer` and `os.PathLike` protocols.
 - Fix incorrect behaviour on Python 3.9 and Python 3.10 that meant that
   calling `isinstance` with `typing_extensions.Concatenate[...]` or
   `typing_extensions.Unpack[...]` as the first argument could have a different